feat(cache): Add contract
A cache should allow putting some arbitrary value, which is associated with a hashable key. If there is a key collision, it should return an error.

It should also allow getting the value associated with the supplied hashable key. If there is no such key, it should return an error.